mcp_server/requirements.txt
mcp
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
